+++ grants_tagger/train.py
@@ -10,7 +10,16 @@ import pickle
 import os.path
 import json
 import ast
+import typer
+import logging
+import time
+import yaml
+import configparser
 
+
+logger = logging.getLogger(__name__)
+
+from typing import List, Optional
 from sklearn.metrics import f1_score, classification_report
 
 from grants_tagger.label_binarizer import create_label_binarizer
@@ -18,6 +27,8 @@ from grants_tagger.models.create_model import create_model
 from grants_tagger.utils import load_train_test_data, yield_tags
 
 from tensorflow.random import set_seed
+from grants_tagger.utils import convert_dvc_to_sklearn_params
+
 
 # TODO: Remove when WellcomeML implements setting random_seed inside models
 # replace with param in configs then
@@ -75,29 +86,88 @@ def train(
             model.save(model_path)
 
 
+train_app = typer.Typer()
+
+
+@train_app.command()
+def train_cli(
+    data_path: Optional[Path] = typer.Argument(
+        None, help="path to processed JSON data to be used for training"
+    ),
+    label_binarizer_path: Optional[Path] = typer.Argument(
+        None, help="path to label binarizer"
+    ),
+    model_path: Optional[Path] = typer.Argument(
+        None, help="path to output model.pkl or dir to save model"
+    ),
+    approach: str = typer.Option("tfidf-svm", help="tfidf-svm, scibert, cnn, ..."),
+    parameters: str = typer.Option(
+        None, help="model params in sklearn format e.g. {'svm__kernel: linear'}"
+    ),
+    threshold: float = typer.Option(None, help="threshold to assign a tag"),
+    data_format: str = typer.Option(
+        "list",
+        help="format that will be used when loading the data. One of list,generator",
+    ),
+    sparse_labels: bool = typer.Option(
+        False, help="flat about whether labels should be sparse when binarized"
+    ),
+    cache_path: Optional[Path] = typer.Option(
+        None, help="path to cache data transformartions"
+    ),
+    config: Path = None,
+):
+
+    start = time.time()
+    params_path = os.path.join(os.path.dirname(__file__), "../params.yaml")
+    with open(params_path) as f:
+        params = yaml.safe_load(f)
+
+    # If parameters not provided from user we initialise from DVC
+    if not parameters:
+        parameters = params["train"].get(approach)
+        parameters = convert_dvc_to_sklearn_params(parameters)
+        parameters = str(parameters)
+
+    # Note that config overwrites parameters for backwards compatibility
+    if config:
+        cfg = configparser.ConfigParser(allow_no_value=True)
+        cfg.read(config)
+
+        config_version = cfg["DEFAULT"]["version"]
+        data_path = cfg["data"]["train_data_path"]
+        label_binarizer_path = cfg["model"]["label_binarizer_path"]
+        approach = cfg["model"]["approach"]
+        parameters = cfg["model"]["parameters"]
+        model_path = cfg["model"].get("model_path", None)
+        threshold = cfg["model"].get("threshold", None)
+        if threshold:
+            threshold = float(threshold)
+        data_format = cfg["data"].get("data_format", "list")
+        sparse_labels = cfg["model"].get("sparse_labels", False)
+        if sparse_labels:
+            sparse_labels = bool(sparse_labels)
+        cache_path = cfg["data"].get("cache_path")
+
+    if model_path and os.path.exists(model_path):
+        print(f"{model_path} exists. Remove if you want to rerun.")
+    else:
+        logger.info(parameters)
+        train(
+            data_path,
+            label_binarizer_path,
+            approach,
+            parameters,
+            model_path=model_path,
+            threshold=threshold,
+            data_format=data_format,
+            sparse_labels=sparse_labels,
+            cache_path=cache_path,
+        )
+
+    duration = time.time() - start
+    print(f"Took {duration:.2f} to train")
+
+
 if __name__ == "__main__":
-    # Note that this CLI is purely for SageMaker so it is quite minimal
-    import argparse
-
-    argparser = argparse.ArgumentParser()
-    argparser.add_argument("--data_path", type=Path)
-    argparser.add_argument("--model_path", type=Path)
-    argparser.add_argument("--label_binarizer_path", type=Path)
-    argparser.add_argument("--approach", type=str)
-    argparser.add_argument("--parameters", type=str)
-    argparser.add_argument("--threshold", type=float)
-    argparser.add_argument("--data_format", type=str, default="list")
-    argparser.add_argument("--sparse_labels", type=bool)
-    argparser.add_argument("--cache_path", type=Path)
-    args = argparser.parse_args()
-
-    train(
-        args.data_path,
-        args.label_binarizer_path,
-        args.approach,
-        args.parameters,
-        threshold=args.threshold,
-        model_path=args.model_path,
-        data_format=args.data_format,
-        sparse_labels=args.sparse_labels,
-    )
+    train_app()
