+++ tests/test_cli.py
@@ -12,7 +12,7 @@ from sklearn.svm import SVC
 from typer.testing import CliRunner
 import pytest
 
-from grants_tagger.__main__ import app, convert_dvc_to_sklearn_params
+from grants_tagger.cli import app, convert_dvc_to_sklearn_params
 from grants_tagger.models.mesh_cnn import MeshCNN
 
 runner = CliRunner()
